+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- properly handle when controller is conifgured for celsius temperatures
+
 - refactor characteristic handlers
 
 ## v0.1.1 - 2020-1-1